Fix crash related to OnPlayerAuthentified handler
Due to changes of how actors data is stored I intentionally left out storing refIds, however this turned out to be a cause for crash related to OnPlayerAuthentified handler. Because of how the script currently works this was the easiest fix to do, even though not as elegant.

Also done some refactoring.